% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{.bridge_js}
\alias{.bridge_js}
\title{Return the path to mermaid_bridge.js in the node directory}
\usage{
.bridge_js()
}
\description{
Return the path to mermaid_bridge.js in the node directory
}
\keyword{internal}
